search function is updated
